Use floats, not doubles
Doubles are the same as floats on the arduino, so there's no point in using
them.

+++ CMPS10.cpp
@@ -31,7 +31,7 @@ byte CMPS10::read_i2c(int address, int _register) {
     return Wire.read();
 }
 
-double CMPS10::read_word(int high_address, int low_address) {
+float CMPS10::read_word(int high_address, int low_address) {
     int high = read_i2c(compass_address, high_address) << 8;
     int low = read_i2c(compass_address, low_address);
     return (high + low) / 10.0;
@@ -59,7 +59,7 @@ CMPS10::CMPS10() {
 /*
  * Return the bearing in degrees. 0-359.9 for a full circle.
  */
-double CMPS10::bearing() {
+float CMPS10::bearing() {
     return read_word(COMPASS_DECIMAL_HIGH, COMPASS_DECIMAL_LOW);
 }
 
@@ -86,18 +86,18 @@ int8_t CMPS10::roll() {
     return (int8_t) read_i2c(compass_address, COMPASS_ROLL);
 }
 
-double CMPS10::raw_to_g(int16_t raw) {
+float CMPS10::raw_to_g(int16_t raw) {
     return raw / (65536.0/4);
 }
 
-double CMPS10::acceleration_x() {
+float CMPS10::acceleration_x() {
     return raw_to_g(read_int16_t(16, 17));
 }
 
-double CMPS10::acceleration_y() {
+float CMPS10::acceleration_y() {
     return raw_to_g(read_int16_t(18, 19));
 }
 
-double CMPS10::acceleration_z() {
+float CMPS10::acceleration_z() {
     return raw_to_g(read_int16_t(20, 21));
 }
